Add logging to preimplemented clients where applicable
