+++ CMakeLists.txt
@@ -5,9 +5,7 @@ set(srcs
     src/Arduino_MQTT_Client.cpp
     src/Arduino_ESP32_Updater.cpp
     src/Arduino_ESP8266_Updater.cpp
-    src/Espressif_Updater.cpp
     src/SDCard_Updater.cpp
-    src/Espressif_MQTT_Client.cpp
     src/HashGenerator.cpp
     src/Helper.cpp
     src/OTA_Update_Callback.cpp
